[refactor] 회원가입, 로그인 반복되는 코드 Auth 컴포넌트로 분리
